Fix checkstyle and add javaDoc
