+++ src/main/java/tfifteenfour/clipboard/logic/commands/addcommand/AddSessionCommand.java
@@ -1,16 +1,18 @@
 package tfifteenfour.clipboard.logic.commands.addcommand;
 
+import static java.util.Objects.requireNonNull;
+
 import tfifteenfour.clipboard.logic.CurrentSelection;
 import tfifteenfour.clipboard.logic.PageType;
 import tfifteenfour.clipboard.logic.commands.CommandResult;
 import tfifteenfour.clipboard.logic.commands.exceptions.CommandException;
 import tfifteenfour.clipboard.model.Model;
-import tfifteenfour.clipboard.model.course.Course;
 import tfifteenfour.clipboard.model.course.Group;
 import tfifteenfour.clipboard.model.course.Session;
 
-import static java.util.Objects.requireNonNull;
-
+/**
+ * Adds a new session to the selected group.
+ */
 public class AddSessionCommand extends AddCommand {
     public static final String COMMAND_TYPE_WORD = "session";
     public static final String MESSAGE_USAGE = COMMAND_WORD + " " + COMMAND_TYPE_WORD
@@ -25,10 +27,22 @@ public class AddSessionCommand extends AddCommand {
 
     private final Session sessionToAdd;
 
+    /**
+     * Creates an AddSessionCommand object to add a session.
+     * @param session The session to be added.
+     */
     public AddSessionCommand(Session session) {
         this.sessionToAdd = session;
     }
 
+    /**
+     * Executes the AddSessionCommand to add a session to the selected group.
+     *
+     * @param model The model of the application.
+     * @param currentSelection The current selection of the user.
+     * @return The result of the command execution.
+     * @throws CommandException If the session already exists in the course or if the user is not on the session page.
+     */
     public CommandResult execute(Model model, CurrentSelection currentSelection) throws CommandException {
         requireNonNull(model);
